update: reorg test cases
